+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Semantic change: stereo becomes matching_cost (transition name, module, files and class name). [#170]
 - Merge image on input configuration section. [#169]
 - Enable the use of GraphMachine if graphviz avalaible to generate machine states diagram. [#149]
+- Move find_valid_neighbors function to img_tools. [#184]
 - Renamed json_cheker to check_json. [#182]
 
 ### Fixed